+++ Sources/Extensions/SwiftStdlib/StringExtensions.swift
@@ -338,6 +338,16 @@ public extension String {
     public var isWhitespace: Bool {
         return self.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty
     }
+    #if os(iOS) || os(tvOS)
+    /// SwifterSwift: Check if the given string spelled correctly
+    public var isSpelledCorrectly: Bool {
+        let checker = UITextChecker()
+        let range = NSRange(location: 0, length: self.utf16.count)
+        
+        let misspelledRange = checker.rangeOfMisspelledWord(in: self, range: range, startingAt: 0, wrap: false, language: Locale.preferredLanguages.first ?? "en")
+        return misspelledRange.location == NSNotFound
+    }
+    #endif
 }
 
 // MARK: - Methods
@@ -856,7 +866,28 @@ public extension String {
 			return self + padding[padding.startIndex..<padding.index(padding.startIndex, offsetBy: padLength)]
 		}
 	}
-	
+    
+    /// SwifterSwift: Removes given prefix from the string.
+    ///
+    ///   "Hello, World!".removingPrefix("Hello, ") -> "World!"
+    ///
+    /// - Parameter prefix: Prefix to remove from the string.
+    /// - Returns: The string after prefix removing.
+    public func removingPrefix(_ prefix: String) -> String {
+        guard self.hasPrefix(prefix) else { return self }
+        return String(self.dropFirst(prefix.count))
+    }
+    
+    /// SwifterSwift: Removes given suffix from the string.
+    ///
+    ///   "Hello, World!".removingSuffix(", World!") -> "Hello"
+    ///
+    /// - Parameter suffix: Suffix to remove from the string.
+    /// - Returns: The string after suffix removing.
+    public func removingSuffix(_ suffix: String) -> String {
+        guard self.hasSuffix(suffix) else { return self }
+        return String(self.dropLast(suffix.count))
+    }
 }
 
 // MARK: - Operators
@@ -1014,5 +1045,4 @@ public extension String {
 	public func appendingPathExtension(_ str: String) -> String? {
 		return (self as NSString).appendingPathExtension(str)
 	}
-	
 }